==> workdir_inference/4_test_parallel_inference_with_simple_examples_using_Ray.ipynb <==
# Perform parallel inference on a GWAS 2022 Axioscan7 images

import argparse
import os
import random
import shutil
import time
import warnings
import pickle
import numpy as np
import math
import sys
import copy
import re
import pandas as pd
import matplotlib.pyplot as plt
import json
import cv2
from itertools import compress

import torch
import torch.nn as nn
import detectron2
from detectron2.utils.logger import setup_logger
setup_logger()
from detectron2 import model_zoo
from detectron2.engine import DefaultPredictor,DefaultTrainer,HookBase
from detectron2.config import get_cfg
from detectron2.utils.visualizer import Visualizer,ColorMode,GenericMask
from detectron2.structures import BoxMode
from detectron2.evaluation import COCOEvaluator,inference_on_dataset
from detectron2.data import build_detection_test_loader,DatasetMapper,build_detection_train_loader,MetadataCatalog,DatasetCatalog
import detectron2.data.transforms as T
import detectron2.utils.comm as comm

import ray
import time

import uuid as uuid
from operator import itemgetter
import seaborn as sns

import shapely
import shapely.geometry
from shapely.geometry import Polygon,MultiPolygon,GeometryCollection
from shapely.validation import make_valid
from shapely.geometry import mapping
#import geopandas as gpd

#import imgfileutils as imf
#import segmentation_tools as sgt
from aicsimageio import AICSImage, imread
from skimage import measure, segmentation
from skimage.measure import regionprops
from skimage.color import label2rgb
#import progressbar
from IPython.display import display, HTML
#from MightyMosaic import MightyMosaic

import glob
from PIL import Image
import csv

# setup directory
root = r'/Users/lovely_shufan/'

project_dir = root + r'Dropbox (Edison_Lab@UGA)/AMF/AMF Imaging 2022/0_inference_using_MaskRCNN_2021/'
output_dir = project_dir + r'2_infer_result/GA_GWAS_2022/'

model_dir = root + r'Dropbox (Edison_Lab@UGA)/AMF/AMF Imaging 2021/2_computer_vision/'

data_dir = r'/Volumes/easystore/GWAS 2022/'

blocks = ['Block2/','Block3/','Block8/']

## check available resources on the computer
Total 20 cores
Estimated time to do inference on 1 czi image is 40 min
to finish inference on 337 accessions x 3 untreated blocks x 40 min = 40440 mins = 674 hrs = 28 days on one core or 2 days on 15 cores

RAY_GCE_TPU_ACCELERATOR_ENDPOINT = (
    "http://metadata.google.internal/computeMetadata/"
    "v1/instance/attributes/accelerator-type"
)
RAY_GCE_TPU_HEADERS = {"Metadata-Flavor": "Google"}
import requests
accelerator_type_request = requests.get(
    RAY_GCE_TPU_ACCELERATOR_ENDPOINT,
    headers=RAY_GCE_TPU_HEADERS,
)

print(accelerator_type_request)

ray.init()
ray.available_resources()['CPU']

ray.init(num_cpus=18, ignore_reinit_error=True)

## test parallel computing using Ray

### Trial 1
everything seems great, code is working

@ray.remote
def timer_ray(x):
    time.sleep(1)
    return x

t0 = time.time()
values = [timer_ray.remote(x) for x in range(4)]
print('Time Elapsed:\t{:.4f}'.format(time.time() - t0))
print(values)

print(ray.get(values))

### Trial 2
Based on code below, ray.get()is extremely slow. Need to optimize ray.get() behavior

def busy(i):
    time.sleep(i)
    return i

@ray.remote
def rbusy(i):
    return busy(i)

# no parallel
t0 = time.time()
objs = [(i, busy(1.0)) for i in range(5)]
print('Time Elapsed:\t{:.4f}'.format(time.time() - t0))
for i,obj in list(objs):
    print(f'{i}: {obj}')

# with parallel
t0 = time.time()
objs = [(i, ray.get(rbusy.remote(1.0)), time.time() - t0) for i in range(5)]
end = time.time() - t0
print('Time Elapsed:\t{:.4f}'.format(time.time() - t0))

for i,obj,t in list(objs):
    print('%d: %s (time: %6.3f seconds)' % (i, obj, t))

This example recreates the problem below. Immediately calling ray.get() on each object ID is effectively blocking the progress, in another word, waited for each remote call to finish before starting the next task. Now we try to solve this by making ray.get() non-blocking.

t0 = time.time()
ids = [(i, rbusy.remote(1.0), time.time() - t0) for i in range(5)] # only calling object id now
t1 = time.time() - t0
print('Time Elapsed:\t{:.5f}'.format(time.time() - t0))

for i,obj_id,t in list(ids):
    print('%d: %s time: %.5f SECONDS' % (i, obj_id, t))

t0 = time.time()
for i, obj_id, t in list(ids):
    print('%d: (%s) time: %.5f SECONDS' % (i, ray.get(obj_id), time.time() - t0))

print('Time Elapsed:\t{:.5f}'.format(time.time() - t0))

The two step process is much faster than directly calling ray.get() after each task. The reason is that while we are waiting for the first task to finish in the second loop, the other four tasks are running in parallel and would finish almost at the same time. We no longer need to wait for the second, third ... task to finish. Therefore, the later ray.get() does not wait long before the object is ready. Now we will generalize this method to task that requires different running time. 

### Trial 3

times = range(5,30,5) # 5, 10, ...
ids = [rbusy.remote(i) for i in times]  # Each waits progressively longer
for x in list(zip(times, ids)):        # Show a list of ids and the times each one will take 
    print(x)

for i in range(1,40): 
    ready, not_ready = ray.wait(ids, num_returns = 2)
    print('iteration:', i) 
    print('Ready length, values: ', len(ready), ray.get(ready))
    print('Not Ready length:', len(not_ready))
    ids = not_ready
    if not ids:
        break

ids = [rbusy.remote(i) for i in times]  # New set of ids
for i in range(1,40):
    num_returns = 2 if len(ids) >= 2 else len(ids)
    ready, not_ready = ray.wait(ids, num_returns = num_returns)
    print('iteration:', i) 
    print('Ready length, values: ', len(ready), ray.get(ready))
    print('Not Ready length:', len(not_ready))
    ids = not_ready
    if not ids: 
        break

## Calculate AXIOSCAN per-channel mean
1. by randomly choosing 30 scenes
    with 1 core
    with 15 cores
2. by randomly choosing 100 images
    with 15 cores

def findBGRmean(file):
    czi = AICSImage(file)
    scenes = czi.scenes
    num_scene = len(scenes)
    x = random.randint(0, num_scene-1)
    czi.set_scene(x)
    img = czi.get_image_data("YXS", T=0,C=0,Z=0)
    
    return np.array([np.mean(img[:,:,0]),np.mean(img[:,:,1]),np.mean(img[:,:,2])])

@ray.remote
def rfindBGRmean(file):
    czi = AICSImage(file)
    scenes = czi.scenes
    x = random.choice(scenes)
    czi.set_scene(x)
    img = czi.get_image_data("YXS", T=0,C=0,Z=0)
    
    return [np.mean(img[:,:,0]),np.mean(img[:,:,1]),np.mean(img[:,:,2])]

# 64G memory limit
memory_limit = 64 * 1024 * 1024 * 1024
print(memory_limit)

# list all files in block2
all_file = [os.path.join(path, name) 
 for path, dirs, files in os.walk(os.path.join(data_dir, 'Block2/'), topdown=False)
 for name in files
if name.endswith('.czi')]
        

print(len(all_file))
print(all_file[0])

### randomly choose 30 scenes and run with 1 core

# randomly choose 30 scenes
sample30 = random.sample(all_file,30)

# calculate channel means without parallel
n=0
t0 = time.time()
for czi in sampleA:
    n = n+1
    print(n)
    print(findBGRmean(czi))
    print('Time Elapsed:\t{:.4f}'.format(time.time() - t0))
print('Total Time Elapsed:\t{:.4f}'.format(time.time() - t0))

### randomly choose 30 scenes and run with 15 cores

# test to make sure the code is finishing in reasonable amount of time
id30 = [rfindBGRmean.remote(czi) for czi in sample30]
t0 = time.time()
for i in range(0,30):
    ready, not_ready = ray.wait(id30, num_returns = 1)
    print('iteration:', i) 
    print('Ready length, values: ', len(ready), ray.get(ready))
    print('Not Ready length:', len(not_ready))
    id30 = not_ready
    if not ids: 
        break
print('Time Elapsed:\t{:.4f}'.format(time.time() - t0))
#print(np.mean(np.array(ray.get(values30)),axis=0))
#print('Time Elapsed:\t{:.4f}'.format(time.time() - t0))

id30 = [rfindBGRmean.remote(czi) for czi in sample30]
results = []
t0 = time.time()
for i in range(1,40):
    ready, not_ready = ray.wait(id30, num_returns = 1)
    print('iteration:', i) 
    results.append(ray.get(ready))
    id30 = not_ready
    if not id30: 
        break
print('Time Elapsed:\t{:.4f}'.format(time.time() - t0))

block2_means = np.mean(results,axis=0)[0]

diff_block2 = np.subtract(train_means, block2_means)
print(block2_means)
print(diff_block2)

# do the same things above for block 3 and 8
means = []
for block in ['Block2/','Block3/','Block8/']:
    print(block)
    results = []
    all_file = []
    t0 = time.time()
    for path, dirs, files in os.walk(os.path.join(data_dir, block), topdown=False):
        for name in files:
            if name.endswith('.czi'):
                all_file.append(os.path.join(path, name))
    sample30 = random.sample(all_file, 30)
    id30 = [rfindBGRmean.remote(czi) for czi in sample30]
    for i in range(1,40):
        ready, not_ready = ray.wait(id30, num_returns = 1)
        print('iteration:', i) 
        results.append(ray.get(ready))
        id30 = not_ready
        if not id30: 
            break
    block_means = np.mean(results,axis=0)[0].tolist()
    means.append(block_means)
    print('Time Elapsed:\t{:.4f}'.format(time.time() - t0))
print('Total Time Elapsed:\t{:.4f}'.format(time.time() - t0))        

train_means = np.array([145.85450147081923, 152.11404784018066, 161.0139541272665])

### Calculate diff between per channels means for centering 

#print(means)
#print(train_means)
#print(np.subtract(train_means, means))
#diffs = np.subtract(train_means, means)
#diffs = np.reshape(np.append(diff_block2,np.subtract(train_means, means)), (3,3))

diffs = [[10.12746633, 18.06137262, 27.510427], 
         [16.24330208, 23.29101365, 34.06829238], 
         [10.1628206,  16.92133011, 26.18143223]]
print(diffs[0])

## parallel inference step-by-step
1. for loops to read in czi images
2. parallelize at the step of looping by tiles

classes=['root','AMF internal hypha','AMF external hypha','AMF arbuscule','AMF vesicle','AMF spore','others']

cfg = get_cfg() # return default configuration
cfg.merge_from_file(model_zoo.get_config_file("COCO-InstanceSegmentation/mask_rcnn_R_50_FPN_3x.yaml")) # copy config files from open source projects

# training configuration
cfg.DATASETS.TEST=()
cfg.DATALOADER.NUM_WORKERS=2
#cfg.SOLVER.IMS_PER_BATCH=args.batch_size

cfg.MODEL.ROI_HEADS.BATCH_SIZE_PER_IMAGE=128 #Number of regions per image used to train RPN. faster, and good enough for this toy dataset (default: 512)
cfg.MODEL.ROI_HEADS.NUM_CLASSES=len(classes)# (see https://detectron2.readthedocs.io/tutorials/datasets.html#update-the-config-for-new-datasets)
cfg.MODEL.BACKBONE.FREEZE_AT=2
cfg.SEED=1
cfg.AUG_FLAG=1

# inference configuration
cfg.MODEL.ROI_HEADS.SCORE_THRESH_TEST = 0.8  # set threshold for this model
cfg.MODEL.WEIGHTS=os.path.join(model_dir, "Trainset1_model_best.pth") # path to the best model trained
cfg.MODEL.DEVICE='cpu' # use cpu for inference


inf_metadata = MetadataCatalog.get("inference").set(thing_classes=['root','AMF internal hypha','AMF external hypha','AMF arbuscule','AMF vesicle','AMF spore','others'])
predictor = DefaultPredictor(cfg)

diff = [10.12746633, 18.06137262, 27.510427]

diff_expanded = diff[:, np.newaxis, np.newaxis]

def centering2train(diff, img):
    row = img.shape[0]
    col = img.shape[1]
    times = row * col
    for i in range(0,3,1):
        diffarray = np.repeat(diff[i], times, axis = 0)
        diffmx = diffarray.reshape((row,col))
        img[:,:,i] = np.add(img[:,:,i],diffmx)
        img[img > 255] = 255
        img[img < 0] = 0
    return img

def padImg(img, tilex, tiley):
    '''
    :param img:
    :return padded img:
    
    :rtype ndnumpy.array:
    Objective: output a padded image dividle by tile size
    '''
    y = img.shape[0]
    x = img.shape[1]
    pad_top = tiley - (y % tiley)
    pad_lft = tilex - (x % tilex)
    img = cv2.copyMakeBorder(img,pad_top,0,pad_lft,0,cv2.BORDER_CONSTANT,value=[0,0,0])
    
    return img

@ray.remote
def inference(pathtofile, block, diff, predictor):
    blklist = []
    imgidlist = []
    sceneidlist = []
    tileidlist = []
    classlist = []
    confscorelist=[]
    arealist = []

    # read in czi
    czi = AICSImage(pathtofile)
   
    for scene in czi.scenes[0:2]:
        # extract image by scene
        czi.set_scene(scene)
        img = czi.get_image_data("YXS", T=0,C=0,Z=0) # numpy.ndarray
                
        # pad image
        img = padImg(img, 2560, 1920)
    
        # centering
        img = centering2train(diff, img)
                 
        # tiling
        for i in range(0,img.shape[0],1920):
            for j in range(0,img.shape[1],2560):
                xmin = j
                xmax = j + 2560
                ymin = i
                ymax = i + 1920
                tile_id = str(xmin)+"_"+str(ymin)+"_"+str(xmax)+"_"+str(ymax)
                subimg = img[ymin:ymax,xmin:xmax]
                outputs = predictor(subimg)
                        
                #inference outputs
                clasind = outputs['instances'].get('pred_classes')
                allmasks = outputs['instances'].get('pred_masks')
                allscores = outputs['instances'].get('scores')
            
                num_seg = clasind.size()[0]
            
                if num_seg != 0: # only save an entry when the image contains a segmentation
                    blklist = blklist + np.repeat(block[:-1], num_seg).tolist()
                    imgidlist = imgidlist + np.repeat(pathtofile, num_seg).tolist()
                    sceneidlist = sceneidlist + np.repeat(scene, num_seg).tolist()
                    tileidlist = tileidlist + np.repeat(tile_id, num_seg).tolist()
                    confscorelist = confscorelist + allscores.tolist()
                
                    # calculate the area of segmentation
                    v = Visualizer(subimg[:, :, ::-1], MetadataCatalog.get("inference"), scale=1.0)
                    for i in range(0,num_seg,1):
                        #calculate mask area
                        locmask = np.asarray(allmasks[i,:,:])
                        gmask = GenericMask(locmask,v.output.height,v.output.width)
                        if gmask.polygons:
                            mergpolygon = gmask.polygons[0]
                            all_points_x = mergpolygon[::2]
                            all_points_y = mergpolygon[1::2]
                            pgon = Polygon(zip(all_points_x,all_points_y))
                            arealist.append(pgon.area)
                            # class index to class name
                            classlist.append(classes[clasind.tolist()[i]])
                        else:
                            continue  # Skip this iteration

    # export inference result as df
    infresults = pd.DataFrame({
    'block': blklist,
    'filename': imgidlist,
    'scene': sceneidlist,
    'tile': tileidlist,
    'annotations': classlist,
    'area': arealist,
    'confidenceScore': confscorelist})
    
    # delete 
    return infresults

## make a test inference run for only two images and two scenes per image

#test_images = os.listdir(os.path.join(data_dir, 'Block2/'))[0:2]
allpath2block2img = [os.path.join(path,name) for path, dirs, files in os.walk(os.path.join(data_dir,'Block2/'))
for name in files
if name.endswith('.czi')]

print(allpath2block2img[0:2])

ids = [inference.remote(path, 'Block2/', diffs[0], predictor) 
       for path in allpath2block2img[0:2]]

results = list()
t0 = time.time()
for i in range(1,340,1):
    ready, not_ready = ray.wait(ids, num_returns = 1)
    print('working on image:', i) 
    results.append(ray.get(ready))
    ids = not_ready
    if not ids: 
        break
    print('Time Elapsed:\t{:.4f}'.format(time.time() - t0))
print('Total Time Elapsed:\t{:.4f}'.format(time.time() - t0))

print(type(results[0][0]))

print(len(results))
print(len(results[0][0]))
print(len(results[1][0]))

df = results[0][0]
for i in range(1,len(results),1):
    newdf = results[i][0]
    df = pd.concat([df, newdf], axis=0)
    
display(df)
print(len(df))

a = list()
a.append(results[0][0])
a.append(results[1][0])

b = pd.concat(a, ignore_index=True)

print(type(a))
print(type(a[1]))
print(a[0].shape)
print(a[1].shape)
print(type(b))
print(b.shape)

print(a[0])

print(a[1])

## Inference on all block 2 images but only two scenes per image

del df

ids = [inference.remote(os.path.join(path,name), 'Block2/', diffs[0], predictor) 
        for path, dirs, files in os.walk(os.path.join(data_dir, 'Block2/'))
        for name in files
        if name.endswith('.czi')]

results = list() #creates a new dataframe that's empty
t0 = time.time()
for i in range(1,340,1):
    ready, not_ready = ray.wait(ids, num_returns = 1)
    print('working on image:', i) 
    results.append(ray.get(ready))
    ids = not_ready
    if not ids: 
        break
    print('Time Elapsed:\t{:.4f}'.format(time.time() - t0))
print('Total Time Elapsed:\t{:.4f}'.format(time.time() - t0))

#%% Clean-up object store data 
del ids


df = results[0][0]
for i in range(1,len(results),1):
    newdf = results[i][0]
    df = pd.concat([df, newdf], axis=0)
df.to_csv(os.path.join(output_dir,"block2_segmentation.txt"),index=False)

print(len(results))
print(len(allpath2block2img))

There are 344 images under block 2 for unknown reasons, expected for < 337 images 

# run the remaining images in block 2
remaining_blk2 = allpath2block2img[-5:]
print(len(remaining_blk2))
print(remaining_blk2)

ids_remain_blk2 = [inference.remote(imgpath, 'Block2/', diffs[0], predictor) 
        for imgpath in remaining_blk2]

results_remain_blk2 = list() #creates a new dataframe that's empty
t0 = time.time()
for i in range(1,340,1):
    ready, not_ready = ray.wait(ids_remain_blk2, num_returns = 1)
    print('working on image:', i) 
    results_remain_blk2.append(ray.get(ready))
    ids_remain_blk2 = not_ready
    if not ids_remain_blk2: 
        break
    print('Time Elapsed:\t{:.4f}'.format(time.time() - t0))
print('Total Time Elapsed:\t{:.4f}'.format(time.time() - t0))

print(len(results_remain_blk2))
print(results_remain_blk2[5])
print([i for i in range(1,5,1)])

df_remain_blk2 = results_remain_blk2[0][0]
for i in range(1,5,1):
    newdf = results_remain_blk2[i][0]
    df_remain_blk2 = pd.concat([df_remain_blk2, newdf], axis=0)
df_remain_blk2.to_csv(os.path.join(output_dir,"block2_segmentation_part2.txt"),index=False)

# Inference on block 3 images and only two scenes per image

### ray.shutdown()


all_files_blk3 = [os.path.join(path,name) 
        for path, dirs, files in os.walk(os.path.join(data_dir, 'Block3/'))
        for name in files
        if name.endswith('.czi')]

ids_blk3 = [inference.remote(os.path.join(path,name), 'Block3/', diffs[1], predictor) 
        for path, dirs, files in os.walk(os.path.join(data_dir, 'Block3/'))
        for name in files
        if name.endswith('.czi')]

#results_blk3 = list() #creates a new dataframe that's empty
t0 = time.time()
for i in range(1,400,1):
    ready, not_ready = ray.wait(ids_blk3, num_returns = 1)
    print('image:', i) 
    results_blk3.append(ray.get(ready))
    ids_blk3 = not_ready
    if not ids_blk3: 
        break
    print('Time Elapsed:\t{:.4f}'.format(time.time() - t0))
print('Total Time Elapsed:\t{:.4f}'.format(time.time() - t0))

print(len(results_blk3))

print(len(all_files_blk3))
print(all_files_blk3[332])
print(all_files_blk3[331:333])

ray.init(num_cpus=18, ignore_reinit_error=True)

ids_blk3 = [inference.remote(path, 'Block3/', diffs[1], predictor) 
        for path in all_files_blk3[88:333]]

#results_blk3 = list() #creates a new dataframe that's empty
t0 = time.time()
for i in range(1,400,1):
    ready, not_ready = ray.wait(ids_blk3, num_returns = 1)
    print('image:', i) 
    results_blk3.append(ray.get(ready))
    ids_blk3 = not_ready
    if not ids_blk3: 
        break
    print('Time Elapsed:\t{:.4f}'.format(time.time() - t0))
print('Total Time Elapsed:\t{:.4f}'.format(time.time() - t0))

ray memory

allpath2block3img = [os.path.join(path,name) for path, dirs, files in os.walk(os.path.join(data_dir,'Block3/'))
for name in files
if name.endswith('.czi')]

print(len(results_blk3[0]))

df_blk3 = results_blk3[0][0]
for i in range(1,len(results_blk3),1):
    newdf = results_blk3[i][0]
    df_blk3 = pd.concat([df_blk3, newdf], axis=0)
#df.to_csv(os.path.join(output_dir,"block3_segmentation.txt"),index=False)

ids = [inference.remote(os.path.join(path,name), 'Block8/', diffs[2], cfg) 
        for path, dirs, files in os.walk(os.path.join(data_dir, 'Block8/'))
        for name in files
        if name.endswith('.czi')]

results = list() #creates a new dataframe that's empty
t0 = time.time()
for i in range(1,400,1):
    ready, not_ready = ray.wait(ids, num_returns = 1)
    print('working on image:', i) 
    results.append(ray.get(ready))
    ids = not_ready
    if not ids: 
        break
    print('Time Elapsed:\t{:.4f}'.format(time.time() - t0))

df = results[0][0]
for i in range(1,len(results),1):
    newdf = results[i][0]
    df = pd.concat([df, newdf], axis=0)
df.to_csv(os.path.join(output_dir,"block8_segmentation.txt"),index=False)

import math
a = math.nan
print(type(a))
