==> workdir_inference/6_report_v2.Rmd <==
---
title: "Inference on High Throughput Microscope Images"
author: "Shufan Zhang"
date: "2023-02-10"
output: rmarkdown::html_vignette
vignette: >
  %\VignetteIndexEntry{Vignette Title}
  %\VignetteEngine{knitr::rmarkdown}
  %\VignetteEncoding{UTF-8}
header-includes:
  - \usepackage{booktabs}
  - \usepackage{longtable}
  - \usepackage{array}
  - \usepackage{multirow}
  - \usepackage{wrapfig}
  - \usepackage{float}
  - \usepackage{colortbl}
  - \usepackage{pdflscape}
  - \usepackage{tabu}
  - \usepackage{threeparttable}
  - \usepackage{threeparttablex}
  - \usepackage[normalem]{ulem}
  - \usepackage{makecell}
  - \usepackage{xcolor}
---

```{r setup, include = FALSE}
knitr::opts_chunk$set(
  collapse = TRUE,
  comment = "#>",
  echo = FALSE,
  warning = FALSE,
  eval = TRUE,
  warning = FALSE, message = FALSE
)

library(emmeans) 
library(GGally)
library(tidyr)
library(ggplot2)
library(viridis)
library(lattice)
library(dplyr)
library(lme4)
library(lmerTest)
library(sjPlot)
library(sjmisc)
library(sjlabelled)
library(performance) # for icc
library(glmmLasso)
#library(MuMIn) # for mixed linear model r squared value
library(pbkrtest)
library(kableExtra)
options(knitr.table.format = "latex")

load("GWAS_2022_inference_data_wrangling_EDA.RData")

my_theme <- theme(text=element_text(size=16),
        plot.title = element_text(face = "bold",size = rel(1.2), hjust = 0.5),
        axis.text.x = element_text(size = 16),
        axis.text.y = element_text(size = 16),
        axis.title.x = element_text(size = 16),
        axis.title.y = element_text(size = 16),
        axis.ticks=element_blank(),
        axis.line=element_blank(),
        panel.border=element_blank(),
        legend.position="none",
        panel.grid.major=element_line(color='#eeeeee'))


```

# Section 1.1: Initial Exploratory Data Analysis of Inference Result
## Method and data cleaning
The best performing Mask RCNN model from 2021 was applied to ZEISS AxioScan7 images from the three untreated blocks in the Georgia GWAS 2022 experiment. Batch effect between training images and new AxioScan7 images was removed by centering the new images. Inference result was deduplicated. Segmentations with confidence score higher than 80%, size larger than 250 pixels are saved for the downstream analyses.The image resolution of AxioScan 7 microscope is 0.120 micron^2/pixel. Segmentation areas were converted to mm^2. 

## Model performance
```{r histograms, fig.retina = 1.5, fig.height=6, fig.width=9, fig.align = "center"}
ggplot(seg_clean_v3, aes(x=confidenceScore,fill=annotations))+
  geom_histogram(position = 'identity',binwidth = 0.005) +
  scale_fill_viridis(discrete=TRUE) +
  labs(fill="")+
  facet_grid(rows = vars(annotations), scales = "free_y")+
  theme_classic()+
  my_theme

ggplot(seg_clean_v3, aes(x=confidenceScore,fill=annotations))+
  geom_histogram(position = 'identity',binwidth = 0.005) +
  scale_fill_viridis(discrete=TRUE) +
  labs(fill="")+
  facet_grid(rows = vars(annotations), cols = vars(block), scales = "free_y")+
  theme_classic()+
  my_theme

ggplot(seg_clean_v3, aes(y=confidenceScore, x=log(area_mm2), color=annotations))+
  geom_point() +
  scale_color_viridis(discrete=TRUE) +
  labs(fill="")+
  facet_grid(rows = vars(annotations))+
  #ylim(c(0,25))+
  theme_classic()+
  my_theme
```
# Section 1.2: More Exploratory Data Analysis 
## More data cleaning
After examining the confidence score and frequency of spore and vesicle, I decided to exclude the two fungal structures from the following analysis. Percent colonization and count density were calculating using model inferred segmentations of AMF extraradical hyphae, intraradical hyphae and arbuscule.

## Distributions of phenotypes
Percent colonization and count density are the two metrics for sorghum overall AMF colonization  calculated from the filtered inference results. 

```{r distributions, fig.retina = 1.5, fig.height=6, fig.width=9, fig.align = "center"}
ggplot(pheno, aes(y=percentCol,x=block,fill=block))+
  geom_violin(scale = "count") +
  geom_jitter(height = 0, width = 0.1)+
  labs(fill="")+
  theme_classic()+
  my_theme

ggplot(pheno, aes(y=percentCol^0.2,x=block,fill=block))+
  geom_violin(scale = "count") +
  geom_jitter(height = 0, width = 0.1)+
  labs(fill="")+
  theme_classic()+
  my_theme

ggplot(pheno, aes(y=countDensity,x=block,fill=block))+
  geom_violin(scale = "count") +
  geom_jitter(height = 0, width = 0.1)+
  labs(fill="")+
  theme_classic()+
  my_theme

ggplot(pheno, aes(y=countDensity^0.2,x=block,fill=block))+
  geom_violin(scale = "count") +
  geom_jitter(height = 0, width = 0.1)+
  labs(fill="")+
  theme_classic()+
  my_theme

ggplot(pheno, aes(x = countDensity^0.2, y = percentCol^0.2, color = block)) +
  geom_point() +
  labs(fill="")+
  #ylim(c(0,25))+
  theme_classic()+
  my_theme
```


# Section 1.3: Block effect 
Nancy mentioned that AMF colonization is highly variable depending on soil texture, nutrients and other environmental factors. I fitted random intercept models for block and accession effects. Chi-square test and Likelihood ratio test were used to test the significance of models. 

## Percent colonization
Statistical tests result suggested that block to be a highly significant effect. Adding accession into the model did not significantly increase the amount of variance explained. 

```{r model selection overall percent colonization random intercept, echo=TRUE}
pc.empty <- lmer(percentCol^0.2 ~ (1|accession), data = pheno)
pc.empty2 <- lmer(percentCol^0.2 ~ (1|block), data = pheno)
pc.blk <- lmer(percentCol^0.2 ~ (1|accession)+(1|block), data = pheno)

anova(pc.empty, pc.blk, test = "Chisq") 
anova(pc.empty, pc.blk, test = "LRT") # AIC, BIC, deviance 
confint(pc.blk)

anova(pc.empty2, pc.blk, test = "Chisq") 
anova(pc.empty2, pc.blk, test = "LRT") # AIC, BIC, deviance 


```

## Count density
Statistical tests result suggested that block to be a highly significant effect. Adding accession into the model did not significantly increase the amount of variance explained. 

```{r model selection overall count density random intercept, echo=TRUE}
cd.empty <- lmer(countDensity^0.2 ~ (1|accession), data = pheno)
cd.empty2 <- lmer(countDensity^0.2 ~ (1|block), data = pheno)
cd.blk <- lmer(countDensity^0.2 ~ (1|accession)+(1|block), data = pheno)
summary(cd.blk)

anova(cd.empty, cd.blk, test = "Chisq") 
anova(cd.empty, cd.blk, test = "LRT") # AIC, BIC, deviance 
confint(cd.blk)

anova(cd.empty2, cd.blk, test = "Chisq") 
anova(cd.empty2, cd.blk, test = "LRT") # AIC, BIC, deviance 

```

## Block effect at fungal structure level
I didn't fit any model here, but slight modification to the models above can be made to test the significance of the interactions between block and fungal structure frequency and size. Figures below indicated that differences in arbuscule size and hyphal frequency could be driving the block effect.

```{r plot block effect, fig.retina = 1.5, fig.height=6, fig.width=9, fig.align = "center"}
ggplot(fungal_area, aes(y=fungalArea,x=annotations,fill=factor(block, levels = c("Block2", "Block8", "Block10"))))+
  geom_boxplot() +
  labs(fill="")+
  theme_classic()+
  theme(text=element_text(size=16),
        plot.title = element_text(face = "bold",size = rel(1.2), hjust = 0.5),
        axis.text.x = element_text(size = 16,angle = -45, hjust = -0.1, vjust = 1.5),
        axis.text.y = element_text(size = 16),
        axis.title.x = element_blank(),
        axis.title.y = element_text(size = 16),
        axis.ticks=element_blank(),
        axis.line=element_blank(),
        panel.border=element_blank(),
        legend.position="right",
        legend.title = element_blank(),
        panel.grid.major=element_line(color='#eeeeee'))


ggplot(fungal_frq, aes(y=freq,x=annotations,fill=factor(block, levels = c("Block2", "Block8", "Block10"))))+
  geom_boxplot() +
  theme_classic()+
  theme(text=element_text(size=16),
        plot.title = element_text(face = "bold",size = rel(1.2), hjust = 0.5),
        axis.text.x = element_text(size = 16,angle = -45, hjust = -0.1, vjust = 1.5),
        axis.text.y = element_text(size = 16),
        axis.title.x = element_blank(),
        axis.title.y = element_text(size = 16),
        axis.ticks=element_blank(),
        axis.line=element_blank(),
        panel.border=element_blank(),
        legend.position="right",
        legend.title = element_blank(),
        panel.grid.major=element_line(color='#eeeeee'))
```


# Section 1.3: Ranking the 340 sorghum accessions by AMF colonization

We would like to verify the reproducibility of a sorghum ranking across the three untreated blocks. The figures below convinced me that the ranking is highly variable between blocks. I could think of two potential reasons to explain the variability. First, the computer vision model is doing a terrible job at identifying AM fungal structures. It was essentially circling regions in the images that were the most similar to training images. The ranking that we were seeing below was not based on AMF colonization but on the similarity of the image to training dataset. Second, we are truly seeing AMF colonization to vary greatly between blocks. It could be that fertilization was not uniform. Border effect could come into play. A combination of the two reasons above could also be true.  

```{r ranking, fig.retina = 1.5, fig.height=6, fig.width=9, fig.align = "center"}
my_theme45 <- theme(text = element_text(size = 16),
        plot.title = element_text(hjust = 0.5),
        axis.text.x = element_blank(),
        axis.title.y = element_text(margin = margin(r = 20), vjust = -0.25),
        legend.position = "right",legend.title= element_blank())

ggplot(pc_blk_anno, aes(fill=annotations, y=percentCol, x=factor(accession, levels = order_pc_blk2)))+ 
    geom_col()+
    scale_fill_viridis_d()+
    theme_classic()+
    facet_grid(~block)+
    xlab("accessions ordered in descending blk2 percent colonization")+
    ylab("percent colonization")+
    theme(text=element_text(size=16),
          plot.title = element_text(face = "bold",size = rel(1.2), hjust = 0.5),
          axis.text.x = element_blank(),
          axis.text.y = element_text(size = 16),
          axis.title.x = element_text(size = 16),
          axis.title.y = element_text(size = 16),
          axis.ticks=element_blank(),
          axis.line=element_blank(),
          panel.border=element_blank(),
          legend.position="right",
          legend.title = element_blank(),
          panel.grid.major=element_line(color='#eeeeee'))

ggplot(cd_blk_anno, aes(fill=annotations, y=countDensity, x=factor(accession, levels = order_cd_blk2)))+
    geom_col()+
    scale_fill_viridis_d()+
    theme_classic()+
    facet_grid(~block)+
    xlab("accessions ordered in descending blk2 count density")+
    ylab("count density")+
    theme(text=element_text(size=16),
          plot.title = element_text(face = "bold",size = rel(1.2), hjust = 0.5),
          axis.text.x = element_blank(),
          axis.text.y = element_text(size = 16),
          axis.title.x = element_text(size = 16),
          axis.title.y = element_text(size = 16),
          axis.ticks=element_blank(),
          axis.line=element_blank(),
          panel.border=element_blank(),
          legend.position="right",
          legend.title = element_blank(),
          panel.grid.major=element_line(color='#eeeeee'))

ggplot(pc_blk_anno, aes(fill=annotations, y=percentCol, x=factor(accession, levels = order_cd_blk2)))+
    geom_col()+
    scale_fill_viridis_d()+
    theme_classic()+
    facet_grid(~block)+
    xlab("accessions ordered in descending blk2 percent colonization")+
    ylab("count density")+
    theme(text=element_text(size=16),
          plot.title = element_text(face = "bold",size = rel(1.2), hjust = 0.5),
          axis.text.x = element_blank(),
          axis.text.y = element_text(size = 16),
          axis.title.x = element_text(size = 16),
          axis.title.y = element_text(size = 16),
          axis.ticks=element_blank(),
          axis.line=element_blank(),
          panel.border=element_blank(),
          legend.position="right",
          legend.title = element_blank(),
          panel.grid.major=element_line(color='#eeeeee'))


ggplot(cd_blk_anno[cd_blk_anno$block == "Block8" & !cd_blk_anno$annotations %in% c("AMF vesicle", "AMF spore"),], aes(fill=annotations, y=countDensity^0.2, x=factor(accession, levels = order_pc_blk8)))+ 
  geom_bar(position="fill", stat="identity")+
  scale_fill_viridis_d()+
  theme_classic()+
  facet_grid(~block)+
  xlab("accession orderd in descending percent colonization")+
  ylab("count density")+
  theme(text=element_text(size=16),
        plot.title = element_text(face = "bold",size = rel(1.2), hjust = 0.5),
        axis.text.x = element_blank(),
        axis.text.y = element_text(size = 16),
        axis.title.x = element_text(size = 16),
        axis.title.y = element_text(size = 16),
        axis.ticks=element_blank(),
        axis.line=element_blank(),
        panel.border=element_blank(),
        legend.position="right",
        legend.title = element_blank(),
        panel.grid.major=element_line(color='#eeeeee'))
```
